Finish implementation of the Asynchronous Acknowledgement Gerät.
This includes a design change for the data format of the AAG's
packets: All numbers are made relative to the initial value,
and the bitset is continuous instead of in MSB-headed chunks.
This is both easier to code and has higher density, especially
when there are fewer than 249 packets to acknowledge.