Split autogenerated C++-Tcl bridge into 17 files.
This gives faster compilation times on systems with slow memory
transfer (<cough>intel!</cough>) and does not worsen it on others.
Additionally, it means we have less to recompile when a non-global
header is changed, that less memory is required for compilation,
and that recompilation of the bridge parallelises better.