Fix block-Gerät corruption from edge case in concurrent modification.
The BlockGeraet now keeps track of which bytes are undergoing concurrent
modification and always sends updates for these bytes. Without this, data could
get corrupted in sequences such as the following:
  1. The known state has a byte set to 0x00.
  2. Mutation A alters the byte to 0x01.
  3. Mutation B (sent before ACK/NAK of A) alters the byte to 0x00. Since this
     matches the last known state, no update is sent.
  4. The remote host processes mutation A. The byte now has value 0x01.
  5. The remote host processes mutation B. The byte now still has value 0x01
     (since the local host didn't send an update for it).
  6. The remote host ACKs A and B simultaneously. Both hosts now believe
     themselves to be fully synchronised, but the local has a byte value 0x00
     and the remote 0x01.
By tracking the concurrent modifications, when B is created, the local host
can tell that it has no idea what the byte value on the remote host might be,
and thus will always include the byte until no concurrent mutations on it remain.